moved metrics in core as union module
