+++ smart_kit/handlers/handle_respond.py
@@ -7,7 +7,7 @@ import scenarios.logging.logger_constants as log_const
 
 from smart_kit.handlers.handler_base import HandlerBase
 from smart_kit.names.action_params_names import TO_MESSAGE_NAME, SEND_TIMESTAMP
-from smart_kit.utils.monitoring import smart_kit_metrics
+from core.monitoring.monitoring import monitoring
 
 
 class HandlerRespond(HandlerBase):
@@ -44,7 +44,7 @@ class HandlerRespond(HandlerBase):
 
         text_preprocessing_result = TextPreprocessingResult(payload.get("message", {}))
 
-        smart_kit_metrics.counter_incoming(self.app_name, user.message.message_name, self.__class__.__name__, user)
+        monitoring.counter_incoming(self.app_name, user.message.message_name, self.__class__.__name__, user)
 
         params = {
             log_const.KEY_NAME: log_const.NORMALIZED_TEXT_VALUE,
